refactored Jacobi diagonalization tests
